+++ iOSPubSub/PubSub/MessageManager.swift
@@ -16,7 +16,7 @@ public class MessageManager{
     
     let mediator : MessageMediator
     
-    public func register(node: MessageNode){
+    public func register(node: ReceivablePublisher){
         mediator.register(node: node)
     }
 }