+++ sample/sample/Loader.swift
@@ -1,0 +1,15 @@
+//
+//  Loader.swift
+//  SampleKit
+//
+//  Created by sangmin park on 1/31/24.
+//
+
+import Foundation
+import UIKit
+
+@objc public class Loader : NSObject{
+    @objc public static func loadModule(){
+        let sample = SampleKit()
+    }
+}